requirements.txt
numpy
matplotlib
pillow
tensorboardX
torch
torchvision
tqdm